relocate ptdf
