+++ opf/__init__.py
@@ -1,5 +1,5 @@
 from .io import parse_file, export_network
-from .core import build_model, compute_ptdf
+from .core import build_model, compute_ptdf, compute_lodf
 
 __name__ = "pyopf"
 __version__ = "0.2.0"
