+++ opf/core/__init__.py
@@ -1,6 +1,5 @@
 from .func import build_model
-# from .warmstart import setup_warmstart
-
-from .utils import compute_ptdf
+from .ptdf import compute_ptdf
+from .lodf import compute_lodf
